react project에 typescript 적용
