+++ src/frontend/src/App.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div>
+      <p>snowball</p>
+    </div>
+  );
+}
+
+export default App;
